feat: support tray menu screenshot action
