+++ src-tauri/src/cmd/window.rs
@@ -1,4 +1,4 @@
-use tauri::{AppHandle, Manager};
+use tauri::AppHandle;
 
 use crate::window;
 
@@ -10,6 +10,7 @@ pub fn show_preview_window(app: AppHandle, path: String) -> Result<String, Strin
 
 #[tauri::command]
 pub fn hide_preview_window(app: AppHandle) -> Result<(), String> {
+    println!("hide_preview_window");
     window::hide_preview_window(&app);
     Ok(())
 }
