+++ src-tauri/src/platform/mac/mod.rs
@@ -1,0 +1,5 @@
+mod screenshot;
+mod window;
+
+pub use screenshot::*;
+pub use window::*;
