+++ src-tauri/src/platform/window/mod.rs
@@ -1,0 +1,3 @@
+mod window;
+
+pub use window::*;
